ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
shap==0.44.0
joblib==1.3.2
apscheduler==3.10.4
httpx==0.25.2
pydantic==2.5.2
